- Move all core bot logic into bot/:   • main.py (entry point)   • commands.py, conversation.py, location_handlers.py, callbacks.py - Extract service integrations into services/:   • qa_service.py   • locker_service.py - Group shared helpers under utils/:   • message_utils.py - Preserve existing resources.py, locker_hours_loader.py, campus_map_data.py - Update entry point to `python -m uni_ai_chatbot.bot.main`
