+++ Golden_Standard/Dockerfile
@@ -4,6 +4,7 @@ FROM python:3.11
 # Set working directory
 WORKDIR /app
 
+RUN COMPOSE_BAKE=true
 # Install required system dependencies
 RUN apt-get update && apt-get install -y \
     build-essential \
@@ -13,7 +14,7 @@ RUN apt-get update && apt-get install -y \
 COPY --from=ghcr.io/astral-sh/uv:latest /uv /uvx /bin/
 
 # Copy requirements file first to leverage Docker cache
-COPY uv.lock pyproject.toml README.md .
+COPY uv.lock pyproject.toml README.md ./
 
 # Install Python dependencies
 RUN uv sync --no-install-package uni_ai_chatbot
@@ -24,4 +25,4 @@ COPY src ./src
 RUN uv sync
 
 # Set the default command (this will be overridden by docker-compose)
-ENTRYPOINT ["uv", "run", "python", "-m", "uni_ai_chatbot.telegram_bot"]
+ENTRYPOINT ["uv", "run", "python", "-m", "uni_ai_chatbot.bot.main"]
