+++ Golden_Standard/docker-compose.yml
@@ -10,5 +10,5 @@ services:
       SUPABASE_URL:               ${SUPABASE_URL}
       SUPABASE_SERVICE_ROLE_KEY:  ${SUPABASE_SERVICE_ROLE_KEY}
     volumes:
-      - ./src:/app/src
+        - ./:/app
     restart: unless-stopped
