+++ Golden_Standard/pyproject.toml
@@ -15,7 +15,6 @@ dependencies = [
     "beautifulsoup4>=4.11.1",
     "pdfplumber>=0.7.4",
     "psycopg2-binary>=2.9.3",
-    "python-dotenv>=0.21.0",
     "supabase>=0.2.0",
     "python-dotenv>=0.21.0",
     "networkx>=2.8.4",
